+++ src/jsonyx/test/test_jsonyx.py
@@ -1,5 +1,6 @@
 """JSON tests."""
 # TODO(Nice Zombies): test schema
+# TODO(Nice Zombies): test paste_values
 from __future__ import annotations
 
 __all__: list[str] = []